requirements.txt
azure-functions
pandas
openpyxl
xlrd
azure-storage-blob
